Add files via upload
This script takes R,G,B subframe stacks (required) and combines them with the narrowband S,H,O stacks to create a "true color" image. That is, the S,H,and O are combined with their respective broadband colors. However, the combining is not done in the usual way of simply adding a portion of the narrowband to the broadband. A narrowband astro image contains not only the photons received from the emission line itself but from a plethora of other sources--the background. This script estimates the true line content of the narrowband stack and adds only that to the broadband.